Changed `r_avertexnormals` to a `double`

+++ WinQuake/gl_rmain.cpp
@@ -279,7 +279,7 @@ void CGLRenderer::R_DrawSpriteModel (entity_t *e)
 
 #define NUMVERTEXNORMALS	162
 
-float	r_avertexnormals[NUMVERTEXNORMALS][3] = {
+double	r_avertexnormals[NUMVERTEXNORMALS][3] = {
 #include "anorms.h"
 };
